requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.3
six==1.17.0
ta==0.11.0
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
